Refactor import paths for ScalarType, StructuredType
The import paths for the ScalarType and StructuredType have been refactored across multiple files, removing the redundant 'base' path segment. The documentation has also been updated to reflect these changes. In addition, these classes are explicitly exported in the mdio.schemas module.
