+++ src/mdio/schemas/v0/variable.py
@@ -4,10 +4,10 @@
 from pydantic import Field
 from pydantic import create_model
 
+from mdio.schemas import ScalarType
+from mdio.schemas import StructuredType
 from mdio.schemas.base.array import NamedArray
 from mdio.schemas.base.core import model_fields
-from mdio.schemas.base.dtype import ScalarType
-from mdio.schemas.base.dtype import StructuredType
 from mdio.schemas.base.encoding import ChunkGridMetadata
 from mdio.schemas.base.metadata import MetadataContainer
 from mdio.schemas.base.metadata import UserAttributes
